Interpolation API. Step 2
Almost functional:
  - Now you can interpolate, for example an argument, either by
    ${"argName"} or ${$.arg("argName")}.

But subtemplate bindings are no longer working (so we must fix it).

Also, when directly called from the template, API, their return value is
not only the return value of the interpolation function (so this SHOULD
be rewired), but also IT IS FED IN AGAIN as data to interpolate
function.

So we must make API functions:
  1. Capable to do their work by themselves (no more piping to the
     interpolation function).
  2. To return undefined. In order to avoid any reinterpolation.