More consistent behaviour between tag methods for arrays
Now, when either $.arg(), $.literal() or $.include() receive an array as
its first argument, returns each element wrapped in brackets.

This behaviour was originally intended for subtemplates ($.include())
and, by mistake, applied only to $.literal().

Updated tests to match this behaviour.

In addition, from now, if subtemplates provided this way have an "alias"
property, it will be rendered as an "as" clause after it.

So now, having:

tpl1 = {
    alias: "subTemplate1",
    sql: "_sql1_",
}

...and so on, `$.include([tpl1, tple2, ...], bindings)` will be rendered
as `(_sql1_) as subTemplate1, (_sql2_)...).

(Subtemplate test still unimplemented)