Renamed SQLTT_ENGINE env var to SQL_ENGINE
This way is more generic (this variable can be used for other purposes
too but hardly for different purposes so its better to match).

The only possible drawback can be if other applications expect different
string for the same database. But this can be easily fixed implementing
suppor for different engine name "aliases". That is: if SQLTT=postgres,
just internally map to "postgresql".