Accept a filter as second parameter in $.data().
Modified $.data() Tag API method in order to accept a second parameter
which, if defined, will let us to filter undesired elements.

If that parameter is a string or regular expression, it will check every
value to match it.

If it is a callback, it will be called with each value and key of the
original data and every element for which this function returns a false
value will be removed.