More powerful $.entries() approach.
Now its formatStr (fourth aparameter) accepts four different wildcards
(%1, %2, $1 and $2).

The number (1 or 2) addresses which part of the entry is addressed (2
for value and 1 for key -except in arrays where still to value-).

The preffex (% or $) now decides whether this value should be
interpolated as a literal ($.literal()) or as an argument ($.arg()).

Its default value for formatStr has been conveniently changed from '%1 =
%2' to '%1 = $2' preserving its original behaviour when it is not
specified.

$.values() implementation now also uses '$2' instead of '%2' for that
parameter in order to keep its usual behaviour.